implement #90: Simplify window "lining"
Now windows are put on a diagonal line as before but with much simpler
code that cuts some corners.  When a window was moved a way or raised,
the entire line is started again.

There are some slight changes in the documentation.

A build target called `gdb-sandbox` was added to `build/GNUmakefile`
that runs `gdb` in addition to the sandbox.