implement #101: Restructure X11 and merge frame/window files
The X11 management was improved by making more clear what parts of X11
we implement but there is no isolation between the X11 code and the
current code!

The frame_* and window_* (excluding window_list) files were merged since
they are rather tiny and using headea sections makes it still clear what
parts are involved.  And now it is easy to just include the frame.h file
when a function should act on a frame.  This is meant to reduce mental
load.