implement #89,#97: Restructure synchronizing
Now buttons/keys are un/grabbed DIRECTLY when they are un/set.  This
makes the code more efficient in the most common case.

The client list is now only synchronized when it actually changed and
PropModeAppend is used when possible.

The `modifiers` action was removed because it had many unintended side
effects when running commands externally and reloading the
configuration.  Now the user can make an alias for the modifier of their
choice etc.