implement #111: Add groups to parser
`group` can now be used to declare a group of actions.  This can be used
to group together a specific set of actions.  But also to make
binding/relation groups.  There is the convenient `ungroup GROUP` that
removes all bindings/associations a group shallowly created.  For each
relation is simply does an unrelate with the same pattern.  And for each
bind a simple unbind.