implement #84: Individual border settings
Add actions to allow individual windows to set their size and color.
This allows to set a border size to 0 for specific windows or change the
focus color for specific windows.

This is also a goodbye to _MOTIF_WM_HINTS as it is a legacy thing and
it just causes inconsistencies.